requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Pillow==8.3.1
requests==2.26.0
Werkzeug==2.0.1

